refactor(workflows): rename to invariant-miner + parameter-discovery; symmetry pass
