+++ .github/workflows/ci.yml
@@ -224,4 +224,4 @@ jobs:
     steps:
       - uses: actions/checkout@v6
       - name: Check schema versions match Dockerfiles
-        run: python3 scripts/check_schema_versions.py
+        run: python3 scripts/check_discovered_schema_versions.py
