+++ .github/workflows/invariant-miner.yml
@@ -1,12 +1,12 @@
-name: Config Rules Refresh
+name: Invariant Miner — Vendor Gate
 
 # Re-vendors the validation-rules JSON by running scripts/vendor_rules.py
 # inside each engine's Docker container. Observes whether the real library
 # matches the corpus's declared expected_outcome; fails CI on divergence;
 # commits regenerated JSON back to the PR.
 #
-# Sibling of schema-refresh.yml (parameter-discovery). Same shape, same
-# commit-back + labelling + PR-comment pattern, different artifact kind.
+# Sibling of parameter-discovery.yml. Same shape, same commit-back +
+# labelling + PR-comment pattern, different artifact kind.
 #
 # Two-tier CI per .product/designs/invariant-miner-design-2026-04-26.md §6:
 #   - vendor-transformers : ubuntu-latest GH-hosted runner (CPU-safe import)
@@ -18,15 +18,15 @@ name: Config Rules Refresh
 #                           llenergymeasure:tensorrt Docker image
 #                           (TRT-LLM 0.21.0 needs CUDA-aware env)
 #
-# Trigger paths (auto):
-#   - docker/Dockerfile.{engine}        (Renovate-driven image bumps,
-#                                        e.g. Dockerfile.vllm)
-#   - configs/validation_rules/{engine}.yaml  (e.g. vllm.yaml)
+# Trigger paths (auto) — Stage 2 vendor gate, refined per the Phase-B trigger
+# policy: fire only on files whose changes the vendor scripts read or replay.
+# Stage-1 mining inputs (scripts/miners/*.py, refresh_invariant_rules.sh)
+# move to auto-mine.yml.
+#   - docker/Dockerfile.*                   (Renovate-driven image bumps)
+#   - configs/validation_rules/*.yaml       (mined corpora regenerated by
+#                                            auto-mine.yml or hand edits)
 #   - scripts/vendor_rules.py
-#   - scripts/miners/{engine}_miner.py        (covered by scripts/miners/*.py
-#                                              glob: vllm_static_miner.py,
-#                                              vllm_dynamic_miner.py, etc.)
-#   - scripts/_vendor_common.py
+#   - scripts/_invariant_vendor_common.py
 #
 # Trigger (manual): workflow_dispatch with explicit engine choice.
 
@@ -37,9 +37,7 @@ on:
       - "docker/Dockerfile.*"
       - "configs/validation_rules/*.yaml"
       - "scripts/vendor_rules.py"
-      - "scripts/_vendor_common.py"
-      - "scripts/miners/*.py"
-      - "scripts/update_engine_rules.sh"
+      - "scripts/_invariant_vendor_common.py"
   workflow_dispatch:
     inputs:
       engine:
@@ -61,7 +59,7 @@ permissions:
   pull-requests: write
 
 concurrency:
-  group: config-rules-refresh-${{ github.head_ref || github.ref }}
+  group: invariant-miner-${{ github.head_ref || github.ref }}
   cancel-in-progress: true
 
 env:
@@ -143,12 +141,12 @@ jobs:
             "docker/Dockerfile.transformers"
             "configs/validation_rules/transformers.yaml"
             "scripts/vendor_rules.py"
-            "scripts/_vendor_common.py"
+            "scripts/_invariant_vendor_common.py"
             "scripts/miners/transformers_miner.py"
             "scripts/miners/transformers_dynamic_miner.py"
             "scripts/miners/_fixpoint_test.py"
             "scripts/miners/_base.py"
-            "scripts/update_engine_rules.sh"
+            "scripts/refresh_invariant_rules.sh"
           )
           SHA=$(git log -1 --format=%H -- "${PATHS[@]}")
           DATE=$(git log -1 --format=%aI -- "${PATHS[@]}")
@@ -173,7 +171,7 @@ jobs:
         run: |
           mkdir -p /tmp/rules-diff
           set +e
-          uv run python scripts/diff_rules.py \
+          uv run python scripts/diff_validation_rules.py \
             /tmp/rules-diff/transformers.old.json \
             "${VENDORED_DIR}/transformers.json" \
             --out /tmp/rules-diff/transformers.md \
@@ -316,14 +314,14 @@ jobs:
             "docker/Dockerfile.vllm"
             "configs/validation_rules/vllm.yaml"
             "scripts/vendor_rules.py"
-            "scripts/_vendor_common.py"
+            "scripts/_invariant_vendor_common.py"
             "scripts/miners/vllm_static_miner.py"
             "scripts/miners/vllm_dynamic_miner.py"
             "scripts/miners/_fixpoint_test.py"
             "scripts/miners/_base.py"
             "scripts/miners/_pydantic_lift.py"
             "scripts/miners/_msgspec_lift.py"
-            "scripts/update_engine_rules.sh"
+            "scripts/refresh_invariant_rules.sh"
           )
           SHA=$(git log -1 --format=%H -- "${PATHS[@]}")
           DATE=$(git log -1 --format=%aI -- "${PATHS[@]}")
@@ -348,7 +346,7 @@ jobs:
         run: |
           mkdir -p /tmp/rules-diff
           set +e
-          uv run python scripts/diff_rules.py \
+          uv run python scripts/diff_validation_rules.py \
             /tmp/rules-diff/vllm.old.json \
             "${VENDORED_DIR}/vllm.json" \
             --out /tmp/rules-diff/vllm.md \
@@ -504,11 +502,11 @@ jobs:
             "docker/Dockerfile.tensorrt"
             "configs/validation_rules/tensorrt.yaml"
             "scripts/vendor_rules.py"
-            "scripts/_vendor_common.py"
+            "scripts/_invariant_vendor_common.py"
             "scripts/miners/tensorrt_miner.py"
             "scripts/miners/tensorrt_static_miner.py"
             "scripts/miners/_base.py"
-            "scripts/update_engine_rules.sh"
+            "scripts/refresh_invariant_rules.sh"
           )
           SHA=$(git log -1 --format=%H -- "${PATHS[@]}")
           DATE=$(git log -1 --format=%aI -- "${PATHS[@]}")
@@ -565,7 +563,7 @@ jobs:
         run: |
           mkdir -p /tmp/rules-diff
           set +e
-          python3 scripts/diff_rules.py \
+          python3 scripts/diff_validation_rules.py \
             /tmp/rules-diff/tensorrt.old.json \
             "${VENDORED_DIR}/tensorrt.json" \
             --out /tmp/rules-diff/tensorrt.md \
