+++ .github/workflows/parameter-discovery.yml
@@ -1,7 +1,22 @@
-name: Schema Refresh
+name: Parameter Discovery Refresh
 
-# Automatically refreshes vendored engine schemas when Renovate bumps a
-# Dockerfile version. Also supports manual dispatch for ad-hoc refreshes.
+# Automatically refreshes vendored engine schemas (discovered parameter
+# spaces) when Renovate bumps a Dockerfile version. Also supports manual
+# dispatch for ad-hoc refreshes.
+#
+# Sibling of invariant-miner.yml. Same App-token + commit-back + comment +
+# label pattern; different artifact kind (introspected schemas vs mined
+# validation rules).
+#
+# Justified asymmetry — single-job loop, not per-engine matrix:
+#   Invariant-miner has three sibling jobs (vendor-transformers, vendor-vllm,
+#   vendor-tensorrt) because each engine's mining + vendor pipeline differs
+#   (different miner modules, different runner type, transformers/vllm on
+#   ubuntu-latest vs tensorrt on self-hosted GPU). Parameter discovery is
+#   generic — refresh_discovered_schemas.sh dispatches per-engine internally
+#   and the runner type is uniform (self-hosted, since vLLM and TRT-LLM may
+#   require CUDA at import time). One job iterating over engines is simpler
+#   here without losing observability.
 #
 # Runs on self-hosted GPU runner (same as gpu-ci.yml) because vLLM and
 # TensorRT-LLM may require CUDA device access at import time.
@@ -11,6 +26,8 @@ on:
     types: [opened, synchronize]
     paths:
       - "docker/Dockerfile.*"
+      - "scripts/discover_engine_schemas.py"
+      - "scripts/refresh_discovered_schemas.sh"
   workflow_dispatch:
     inputs:
       engine:
@@ -32,23 +49,34 @@ permissions:
   pull-requests: write
 
 concurrency:
-  group: schema-refresh-${{ github.head_ref || github.ref }}
+  group: parameter-discovery-${{ github.head_ref || github.ref }}
   cancel-in-progress: true
 
 jobs:
   refresh:
-    # Auto-trigger: only for Renovate PRs. Manual dispatch: always run.
-    if: >-
-      github.event_name == 'workflow_dispatch'
-      || github.event.pull_request.user.login == 'renovate[bot]'
     runs-on: self-hosted
     timeout-minutes: 90
     steps:
+      - name: Mint llem-ci-bot App token
+        # Short-lived (~1 hour) GitHub App token. Events it produces
+        # (commits, PR comments, label edits) are NOT subject to the
+        # default GITHUB_TOKEN recursive-workflow-protection gate, so
+        # the commit-back below can legitimately trigger subsequent
+        # CI runs. Skipped on fork PRs (secrets are not exposed there)
+        # so the workflow still runs read-only for external contributors.
+        id: app-token
+        if: github.event.pull_request.head.repo.full_name == github.repository || github.event_name == 'workflow_dispatch'
+        uses: actions/create-github-app-token@v1
+        with:
+          app-id: ${{ secrets.APP_ID }}
+          private-key: ${{ secrets.APP_PRIVATE_KEY }}
+
       - name: Checkout PR branch
         uses: actions/checkout@v6
         with:
           ref: ${{ github.head_ref || github.ref }}
           fetch-depth: 0
+          token: ${{ steps.app-token.outputs.token || secrets.GITHUB_TOKEN }}
 
       - name: Determine engines to refresh
         id: engines
@@ -86,7 +114,7 @@ jobs:
             fi
 
             # Run discovery.
-            ./scripts/update_engine_schema.sh "$engine" || {
+            ./scripts/refresh_discovered_schemas.sh "$engine" || {
               echo "::error::Discovery failed for $engine"
               exit 1
             }
@@ -108,7 +136,7 @@ jobs:
               continue
             fi
 
-            OUTPUT=$(python3 scripts/diff_schemas.py "$OLD" "$SCHEMA" 2>/dev/null) || true
+            OUTPUT=$(python3 scripts/diff_discovered_schemas.py "$OLD" "$SCHEMA" 2>/dev/null) || true
             read -r BREAKING SUMMARY < <(echo "$OUTPUT" | python3 -c "
           import sys, json
           d = json.load(sys.stdin)
@@ -126,9 +154,12 @@ jobs:
           printf '%b' "$DIFF_BODY" > /tmp/diff_body.md
 
       - name: Commit schema updates
+        # Skipped for forks (App token isn't minted there) and when no App
+        # token was provided.
+        if: steps.app-token.outputs.token != '' && github.event_name == 'pull_request' && github.event.pull_request.head.repo.full_name == github.repository
         run: |
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
+          git config user.name  "llem-ci-bot[bot]"
+          git config user.email "${{ secrets.APP_ID }}+llem-ci-bot[bot]@users.noreply.github.com"
 
           git add src/llenergymeasure/config/discovered_schemas/*.json
 
@@ -139,12 +170,15 @@ jobs:
 
           ENGINES="${{ steps.engines.outputs.engines }}"
           git commit -m "chore(config): refresh ${ENGINES} schema(s)"
-          git push
+          # --force-with-lease guards against clobbering a concurrent push
+          # from the PR author. On stale-ref failure the workflow exits
+          # non-zero and a rerun picks up the latest state.
+          git push --force-with-lease
 
       - name: Post diff comment
         if: github.event_name == 'pull_request' || inputs.pr_number != ''
         env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          GH_TOKEN: ${{ steps.app-token.outputs.token || secrets.GITHUB_TOKEN }}
         run: |
           PR_NUMBER="${{ github.event.pull_request.number || inputs.pr_number }}"
           if [[ -z "$PR_NUMBER" ]]; then
@@ -158,7 +192,7 @@ jobs:
       - name: Apply label
         if: github.event_name == 'pull_request' || inputs.pr_number != ''
         env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          GH_TOKEN: ${{ steps.app-token.outputs.token || secrets.GITHUB_TOKEN }}
         run: |
           PR_NUMBER="${{ github.event.pull_request.number || inputs.pr_number }}"
           if [[ -z "$PR_NUMBER" ]]; then
