+++ scripts/_invariant_vendor_common.py
