+++ tests/unit/scripts/test_diff_validation_rules.py
@@ -1,4 +1,4 @@
-"""Tests for :mod:`scripts.diff_rules`."""
+"""Tests for :mod:`scripts.diff_validation_rules`."""
 
 from __future__ import annotations
 
@@ -12,9 +12,9 @@ _PROJECT_ROOT = Path(__file__).resolve().parents[3]
 if str(_PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(_PROJECT_ROOT))
 
-from scripts import diff_rules  # noqa: E402
+from scripts import diff_validation_rules as diff_rules  # noqa: E402
 
-SCRIPT = _PROJECT_ROOT / "scripts" / "diff_rules.py"
+SCRIPT = _PROJECT_ROOT / "scripts" / "diff_validation_rules.py"
 
 
 def _envelope(cases: list[dict[str, Any]], **meta: Any) -> dict[str, Any]:
